+++ os-network/bench/rc_conn_server/src/native/kernel_helper.c
@@ -3,8 +3,8 @@
 #define BUF_LENGTH 256
 #define DEFAULT_PERMISSION S_IRUSR|S_IWUSR|S_IRGRP|S_IROTH
 
-long nic_count = 2;
-module_param(nic_count, long, DEFAULT_PERMISSION);
+uint nic_count = 2;
+module_param(nic_count, uint, DEFAULT_PERMISSION);
 
-long service_id_base = 50;
-module_param(service_id_base, long, DEFAULT_PERMISSION);
+uint service_id_base = 50;
+module_param(service_id_base, uint, DEFAULT_PERMISSION);
